fix: cancel stopping countdown on server_stop
HibernateR and similar plugins stop the server daemon thread but keep
MCDR daemon keep running. The countdown may try to stop the server again
after the server daemon is exitted and the pesudo server is listening,
which raise a warning from MCDR that a server cannot be killed twice.
Mark up the cancel sign for the countdown can fix this minor issue.