fix: trigger server-saved countdown only if server is stopping
Currently this countdown will trigger once server saving completed no
mather under what condition. This may conflict with backup plugins/mod
such as PrimeBackup. In this situation, server only saves the world but
does not wish to stop. Checking that provided the server is in a
stopping sequence to trigger the countdown addresses this issue.